Update cpp_Test_1-4.py
Compared to the original file Test_1-4.cpp which contains 232 lines (ignoring the space from 233), and the previous version of this file which has 35 lines of code (ignoring the space from 36), this program cuts off a lot of lines by compressing series of strings into a single line of list; although, obviously, this style is inefficient and is much exhausting than the previous formats, unless the programmer's already used on using "\n" and "\t" within the single line. Also, I'll try to make an improvised version of this in the future. Lastly, yes, I'm lazy enough to finish this off so yeahh