Add files via upload
Compiled and compressed files of Spacebar (supposedly) forum.
I'm planning to discontinue it and make a new project - the website and its database just suck.

You can check it out: http://spacebar.epizy.com

Don't worry, the only info that's going to be read (not collected) is your IP address, which is very not trustworthy. You can check out the PHP to get what I mean.