Update: Ordinal Numbers/nth's String Emphasis
This is my very first time since last time I had my hands on Python, to program and update the contents without relying from any tutorials, and this made me happy.

It works correctly, meaning it shows if within the nth terms it's the 1st or 2nd and so on, still, I am afraid that beyond 20th term the string of succeeding ordinal numbers will still be displayed with 'th', e.g. 21th, 22th, and so on, which is wrong.

I've yet to fix this one lol.